Fixes /etc/network/interfaces.
Fixes non-failing error whereby the us mirror was used, instead of our local.

Removes lines that are not compatible with ubuntu's implementation of
kickstart.

Calls curl in the target chroot , as 14.04 does not have curl on the
installation media.

Introduces configurable credentials as I could not find them in the config.yml.